Added authorization to the site

+++ app/controllers/todos_controller.rb
@@ -3,18 +3,24 @@ respond_to :html, :js
 	def index
 		@todos = Todo.all
 		@todo = Todo.new
+		authorize @todos
+		
 	end
 
 	def new
 		@todo = Todo.new
+		authorize @todo
 	end
 
 	def show
 		@todo = Todo.find params[:id]
+		authorize @todo
 	end
 
 	def create
 		@todo = Todo.new(todo_params)
+		authorize @todo
+
 		if @todo.save
 			flash[:notice] = "Your new TODO was saved."
 			redirect_to todos_path
@@ -26,6 +32,7 @@ respond_to :html, :js
 
 	def destroy
 		@todo = Todo.find(params[:id])
+		authorize @todo
 		desc = @todo.description
 
 		if @todo.destroy
